pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]


[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.35.64"
pyspark = {extras = ["sql"], version = "^3.5.3"}
awsglue-dev = "^2021.12.30"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
coverage = "^7.6.7"
setuptools = "^75.6.0"


[tool.poetry.plugins.dotenv]
ignore = "false"
location = ".env"



[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-mock = "^3.14.0"

[tool.pytest.ini_options]
testpaths = ["."]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
